requirements.txt
# ElectraLens - Complete Project Requirements
# Includes both frontend and backend dependencies

# Frontend Dependencies
streamlit==1.29.0
requests==2.31.0
pandas==2.1.4
plotly==5.18.0

# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.44
pydantic==2.5.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
slowapi==0.1.9
python-jose[cryptography]==3.5.0

# Database Support
psycopg2-binary==2.9.11

# Common Dependencies
python-dotenv==1.0.0
